QOL changes.
The responsive design is more optimized and now properly displays on mobile displays.

MUI is completely deleted from this project as it was causing issues.

On the game-over screen, The user can now click on the "Restart Game" button to reset the game. This eliminates the reason to refresh the page when failed the game to reset the game.